pyproject.toml
[tool.poetry]
name = "eink-train-display"
version = "0.1.0"
description = "Waveshare eink display program"
authors = ["Jonathan Nakandala <jonathannakandala@gmail.com>"]
readme = "readme.md"
packages = [{ include = "sources" }]

[tool.poetry.dependencies]
python = "^3.9"
zeep = "3.4.0"
spidev = "3.4"
"RPi.GPIO" = "0.7.1"
Pillow = "8.3.2"
svglib = "0.9.4"
pydantic = "^1.10.9"
structlog = "^23.1.0"
httpx = "^0.24.1"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pylint = "^2.17.4"
pytest = "^7.3.2"
pylint-pydantic = "^0.1.8"
pytest-httpx = "^0.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.MASTER]
load-plugins = "pylint_pydantic"

[tool.pylint.'MESSAGES CONTROL']
extension-pkg-whitelist = "pydantic"
